PayCustomer #04

+++ src/br/com/gda/resource/StoreResource.java
@@ -104,7 +104,7 @@ public class StoreResource {
 	@DELETE
 	@Path(DELETE_STORE)
 	public Response deleteStore(@HeaderParam("codOwner") @DefaultValue("-1") long codOwner, 
-			                    @HeaderParam("codStore") @DefaultValue("-1") int codStore) {
+			                    @HeaderParam("codStore") @DefaultValue("-1") long codStore) {
 		StoreInfo recordInfo = new StoreInfo();
 		recordInfo.codOwner = codOwner;
 		recordInfo.codStore = codStore;
